chore: move viewportToCanvas and canvasToViewport in viewport.ts
